+++ codec-mqtt/src/main/java/io/netty/handler/codec/mqtt/MqttDecoder.java
@@ -27,8 +27,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static io.netty.handler.codec.mqtt.MqttCodecUtil.isValidClientId;
-import static io.netty.handler.codec.mqtt.MqttCodecUtil.isValidMessageId;
-import static io.netty.handler.codec.mqtt.MqttCodecUtil.isValidPublishTopicName;
 import static io.netty.handler.codec.mqtt.MqttCodecUtil.resetUnusedFields;
 import static io.netty.handler.codec.mqtt.MqttCodecUtil.validateFixedHeader;
 
@@ -59,13 +57,16 @@ public final class MqttDecoder extends ReplayingDecoder<DecoderState> {
 
     private final int maxBytesInMessage;
 
-    public MqttDecoder() {
-      this(DEFAULT_MAX_BYTES_IN_MESSAGE);
+    private final IVariableHeaderDecoder headerDecoder;
+
+    public MqttDecoder(IVariableHeaderDecoder headerDecoder) {
+      this(DEFAULT_MAX_BYTES_IN_MESSAGE, headerDecoder);
     }
 
-    public MqttDecoder(int maxBytesInMessage) {
+    public MqttDecoder(int maxBytesInMessage, IVariableHeaderDecoder headerDecoder) {
         super(DecoderState.READ_FIXED_HEADER);
         this.maxBytesInMessage = maxBytesInMessage;
+        this.headerDecoder = headerDecoder;
     }
 
     @Override
@@ -85,7 +86,7 @@ public final class MqttDecoder extends ReplayingDecoder<DecoderState> {
                 if (bytesRemainingInVariablePart > maxBytesInMessage) {
                     throw new DecoderException("too large message: " + bytesRemainingInVariablePart + " bytes");
                 }
-                final Result<?> decodedVariableHeader = decodeVariableHeader(buffer, mqttFixedHeader);
+                final Result<?> decodedVariableHeader = headerDecoder.decodeVariableHeader(buffer, mqttFixedHeader);
                 variableHeader = decodedVariableHeader.value;
                 bytesRemainingInVariablePart -= decodedVariableHeader.numberOfBytesConsumed;
                 checkpoint(DecoderState.READ_PAYLOAD);
@@ -160,7 +161,7 @@ public final class MqttDecoder extends ReplayingDecoder<DecoderState> {
         return validateFixedHeader(resetUnusedFields(decodedFixedHeader));
     }
 
-    private static Result<Integer> decodeVariableByteInteger(ByteBuf buffer) {
+    static Result<Integer> decodeVariableByteInteger(ByteBuf buffer) {
         int remainingLength = 0;
         int multiplier = 1;
         short digit;
@@ -179,206 +180,206 @@ public final class MqttDecoder extends ReplayingDecoder<DecoderState> {
         return new Result<Integer>(remainingLength, loops);
     }
 
-    /**
-     * Decodes the variable header (if any)
-     * @param buffer the buffer to decode from
-     * @param mqttFixedHeader MqttFixedHeader of the same message
-     * @return the variable header
-     */
-    private static Result<?> decodeVariableHeader(ByteBuf buffer, MqttFixedHeader mqttFixedHeader) {
-        switch (mqttFixedHeader.messageType()) {
-            case CONNECT:
-                return decodeConnectionVariableHeader(buffer);
+//    /**
+//     * Decodes the variable header (if any)
+//     * @param buffer the buffer to decode from
+//     * @param mqttFixedHeader MqttFixedHeader of the same message
+//     * @return the variable header
+//     */
+//    private static Result<?> decodeVariableHeader(ByteBuf buffer, MqttFixedHeader mqttFixedHeader) {
+//        switch (mqttFixedHeader.messageType()) {
+//            case CONNECT:
+//                return decodeConnectionVariableHeader(buffer);
+//
+//            case CONNACK:
+//                return decodeConnAckVariableHeader(buffer);
+//
+//            case SUBSCRIBE:
+//            case UNSUBSCRIBE:
+//            case SUBACK:
+//            case UNSUBACK:
+//            case PUBACK:
+//            case PUBREC:
+//            case PUBCOMP:
+//            case PUBREL:
+//                return decodeMessageIdVariableHeader(buffer);
+//
+//            case PUBLISH:
+//                return decodePublishVariableHeader(buffer, mqttFixedHeader);
+//
+//            case PINGREQ:
+//            case PINGRESP:
+//            case DISCONNECT:
+//                // Empty variable header
+//                return new Result<Object>(null, 0);
+//        }
+//        return new Result<Object>(null, 0); //should never reach here
+//    }
 
-            case CONNACK:
-                return decodeConnAckVariableHeader(buffer);
+//    private static Result<MqttConnectVariableHeader> decodeConnectionVariableHeader(ByteBuf buffer) {
+//        final Result<String> protoString = decodeString(buffer);
+//        int numberOfBytesConsumed = protoString.numberOfBytesConsumed;
+//
+//        final byte protocolLevel = buffer.readByte();
+//        numberOfBytesConsumed += 1;
+//
+//        final MqttVersion mqttVersion = MqttVersion.fromProtocolNameAndLevel(protoString.value, protocolLevel);
+//
+//        final int b1 = buffer.readUnsignedByte();
+//        numberOfBytesConsumed += 1;
+//
+//        final Result<Integer> keepAlive = decodeMsbLsb(buffer);
+//        numberOfBytesConsumed += keepAlive.numberOfBytesConsumed;
+//
+//        final boolean hasUserName = (b1 & 0x80) == 0x80;
+//        final boolean hasPassword = (b1 & 0x40) == 0x40;
+//        final boolean willRetain = (b1 & 0x20) == 0x20;
+//        final int willQos = (b1 & 0x18) >> 3;
+//        final boolean willFlag = (b1 & 0x04) == 0x04;
+//        final boolean cleanSession = (b1 & 0x02) == 0x02;
+//        if (mqttVersion == MqttVersion.MQTT_3_1_1) {
+//            final boolean zeroReservedFlag = (b1 & 0x01) == 0x0;
+//            if (!zeroReservedFlag) {
+//                // MQTT v3.1.1: The Server MUST validate that the reserved flag in the CONNECT Control Packet is
+//                // set to zero and disconnect the Client if it is not zero.
+//                // See http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc385349230
+//                throw new DecoderException("non-zero reserved flag");
+//            }
+//        }
+//
+//        final Result<MqttProperties> properties;
+//        if (mqttVersion == MqttVersion.MQTT_5) {
+//            properties = decodeProperties(buffer);
+//        } else {
+//            properties = new Result<MqttProperties>(MqttProperties.NO_PROPERTIES, 0);
+//        }
+//        numberOfBytesConsumed += properties.numberOfBytesConsumed;
+//
+//        final MqttConnectVariableHeader mqttConnectVariableHeader = new MqttConnectVariableHeader(
+//                mqttVersion.protocolName(),
+//                mqttVersion.protocolLevel(),
+//                hasUserName,
+//                hasPassword,
+//                willRetain,
+//                willQos,
+//                willFlag,
+//                cleanSession,
+//                keepAlive.value,
+//                properties.value);
+//        return new Result<MqttConnectVariableHeader>(mqttConnectVariableHeader, numberOfBytesConsumed);
+//    }
 
-            case SUBSCRIBE:
-            case UNSUBSCRIBE:
-            case SUBACK:
-            case UNSUBACK:
-            case PUBACK:
-            case PUBREC:
-            case PUBCOMP:
-            case PUBREL:
-                return decodeMessageIdVariableHeader(buffer);
+//    private static Result<MqttProperties> decodeProperties(ByteBuf buffer) {
+//        final Result<Integer> propertiesLength = decodeVariableByteInteger(buffer);
+//        int totalPropertiesLength = propertiesLength.value;
+//        int numberOfBytesConsumed = propertiesLength.numberOfBytesConsumed;
+//
+//        MqttProperties decodedProperties = new MqttProperties();
+//        while (numberOfBytesConsumed < totalPropertiesLength) {
+//            Result<Integer> propertyId = decodeVariableByteInteger(buffer);
+//            numberOfBytesConsumed += propertyId.numberOfBytesConsumed;
+//
+//            switch (propertyId.value) {
+//                case 0x01: // Payload Format Indicator => Byte
+//                case 0x17: // Request Problem Information
+//                case 0x19: // Request Response Information
+//                case 0x24: // Maximum QoS
+//                case 0x25: // Retain Available
+//                case 0x28: // Wildcard Subscription Available
+//                case 0x29: // Subscription Identifier Available
+//                case 0x2A: // Shared Subscription Available
+//                    final int b1 = buffer.readUnsignedByte();
+//                    numberOfBytesConsumed ++;
+//                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, b1));
+//                    break;
+//                case 0x13: // Server Keep Alive => Two Byte Integer
+//                case 0x21: // Receive Maximum
+//                case 0x22: // Topic Alias Maximum
+//                case 0x23: // Topic Alias
+//                    final Result<Integer> int2BytesResult = decodeMsbLsb(buffer);
+//                    numberOfBytesConsumed += int2BytesResult.numberOfBytesConsumed;
+//                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, int2BytesResult.value));
+//                    break;
+//                case 0x02: // Publication Expiry Interval => Four Byte Integer
+//                case 0x11: // Session Expiry Interval
+//                case 0x18: // Will Delay Interval
+//                case 0x27: // Maximum Packet Size
+//                    final Result<Integer> int4BytesResult = decode4bytesInteger(buffer);
+//                    numberOfBytesConsumed += int4BytesResult.numberOfBytesConsumed;
+//                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, int4BytesResult.value));
+//                    break;
+//                case 0x0B: // Subscription Identifier => Variable Byte Integer
+//                    Result<Integer> vbIntegerResult = decodeVariableByteInteger(buffer);
+//                    numberOfBytesConsumed += vbIntegerResult.numberOfBytesConsumed;
+//                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, vbIntegerResult.value));
+//                    break;
+//                case 0x03: // Content Type => UTF-8 Encoded String
+//                case 0x08: // Response Topic
+//                case 0x12: // Assigned Client Identifier
+//                case 0x15: // Authentication Method
+//                case 0x1A: // Response Information
+//                case 0x1C: // Server Reference
+//                case 0x1F: // Reason String
+//                case 0x26: // User Property
+//                    final Result<String> stringResult = decodeString(buffer);
+//                    numberOfBytesConsumed += stringResult.numberOfBytesConsumed;
+//                    decodedProperties.add(new MqttProperties.StringProperty(propertyId.value, stringResult.value));
+//                    break;
+//                case 0x09: // Correlation Data => Binary Data
+//                case 0x16: // Authentication Data
+//                    final Result<byte[]> binaryDataResult = decodeByteArray(buffer);
+//                    numberOfBytesConsumed += binaryDataResult.numberOfBytesConsumed;
+//                    decodedProperties.add(new MqttProperties.BinaryProperty(propertyId.value, binaryDataResult.value));
+//                    break;
+//            }
+//        }
+//
+//        return new Result<MqttProperties>(decodedProperties, numberOfBytesConsumed);
+//    }
 
-            case PUBLISH:
-                return decodePublishVariableHeader(buffer, mqttFixedHeader);
-
-            case PINGREQ:
-            case PINGRESP:
-            case DISCONNECT:
-                // Empty variable header
-                return new Result<Object>(null, 0);
-        }
-        return new Result<Object>(null, 0); //should never reach here
-    }
-
-    private static Result<MqttConnectVariableHeader> decodeConnectionVariableHeader(ByteBuf buffer) {
-        final Result<String> protoString = decodeString(buffer);
-        int numberOfBytesConsumed = protoString.numberOfBytesConsumed;
-
-        final byte protocolLevel = buffer.readByte();
-        numberOfBytesConsumed += 1;
-
-        final MqttVersion mqttVersion = MqttVersion.fromProtocolNameAndLevel(protoString.value, protocolLevel);
-
-        final int b1 = buffer.readUnsignedByte();
-        numberOfBytesConsumed += 1;
-
-        final Result<Integer> keepAlive = decodeMsbLsb(buffer);
-        numberOfBytesConsumed += keepAlive.numberOfBytesConsumed;
-
-        final boolean hasUserName = (b1 & 0x80) == 0x80;
-        final boolean hasPassword = (b1 & 0x40) == 0x40;
-        final boolean willRetain = (b1 & 0x20) == 0x20;
-        final int willQos = (b1 & 0x18) >> 3;
-        final boolean willFlag = (b1 & 0x04) == 0x04;
-        final boolean cleanSession = (b1 & 0x02) == 0x02;
-        if (mqttVersion == MqttVersion.MQTT_3_1_1) {
-            final boolean zeroReservedFlag = (b1 & 0x01) == 0x0;
-            if (!zeroReservedFlag) {
-                // MQTT v3.1.1: The Server MUST validate that the reserved flag in the CONNECT Control Packet is
-                // set to zero and disconnect the Client if it is not zero.
-                // See http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc385349230
-                throw new DecoderException("non-zero reserved flag");
-            }
-        }
-
-        final Result<MqttProperties> properties;
-        if (mqttVersion == MqttVersion.MQTT_5) {
-            properties = decodeProperties(buffer);
-        } else {
-            properties = new Result<MqttProperties>(MqttProperties.NO_PROPERTIES, 0);
-        }
-        numberOfBytesConsumed += properties.numberOfBytesConsumed;
-
-        final MqttConnectVariableHeader mqttConnectVariableHeader = new MqttConnectVariableHeader(
-                mqttVersion.protocolName(),
-                mqttVersion.protocolLevel(),
-                hasUserName,
-                hasPassword,
-                willRetain,
-                willQos,
-                willFlag,
-                cleanSession,
-                keepAlive.value,
-                properties.value);
-        return new Result<MqttConnectVariableHeader>(mqttConnectVariableHeader, numberOfBytesConsumed);
-    }
-
-    private static Result<MqttProperties> decodeProperties(ByteBuf buffer) {
-        final Result<Integer> propertiesLength = decodeVariableByteInteger(buffer);
-        int totalPropertiesLength = propertiesLength.value;
-        int numberOfBytesConsumed = propertiesLength.numberOfBytesConsumed;
-
-        MqttProperties decodedProperties = new MqttProperties();
-        while (numberOfBytesConsumed < totalPropertiesLength) {
-            Result<Integer> propertyId = decodeVariableByteInteger(buffer);
-            numberOfBytesConsumed += propertyId.numberOfBytesConsumed;
-
-            switch (propertyId.value) {
-                case 0x01: // Payload Format Indicator => Byte
-                case 0x17: // Request Problem Information
-                case 0x19: // Request Response Information
-                case 0x24: // Maximum QoS
-                case 0x25: // Retain Available
-                case 0x28: // Wildcard Subscription Available
-                case 0x29: // Subscription Identifier Available
-                case 0x2A: // Shared Subscription Available
-                    final int b1 = buffer.readUnsignedByte();
-                    numberOfBytesConsumed ++;
-                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, b1));
-                    break;
-                case 0x13: // Server Keep Alive => Two Byte Integer
-                case 0x21: // Receive Maximum
-                case 0x22: // Topic Alias Maximum
-                case 0x23: // Topic Alias
-                    final Result<Integer> int2BytesResult = decodeMsbLsb(buffer);
-                    numberOfBytesConsumed += int2BytesResult.numberOfBytesConsumed;
-                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, int2BytesResult.value));
-                    break;
-                case 0x02: // Publication Expiry Interval => Four Byte Integer
-                case 0x11: // Session Expiry Interval
-                case 0x18: // Will Delay Interval
-                case 0x27: // Maximum Packet Size
-                    final Result<Integer> int4BytesResult = decode4bytesInteger(buffer);
-                    numberOfBytesConsumed += int4BytesResult.numberOfBytesConsumed;
-                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, int4BytesResult.value));
-                    break;
-                case 0x0B: // Subscription Identifier => Variable Byte Integer
-                    Result<Integer> vbIntegerResult = decodeVariableByteInteger(buffer);
-                    numberOfBytesConsumed += vbIntegerResult.numberOfBytesConsumed;
-                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, vbIntegerResult.value));
-                    break;
-                case 0x03: // Content Type => UTF-8 Encoded String
-                case 0x08: // Response Topic
-                case 0x12: // Assigned Client Identifier
-                case 0x15: // Authentication Method
-                case 0x1A: // Response Information
-                case 0x1C: // Server Reference
-                case 0x1F: // Reason String
-                case 0x26: // User Property
-                    final Result<String> stringResult = decodeString(buffer);
-                    numberOfBytesConsumed += stringResult.numberOfBytesConsumed;
-                    decodedProperties.add(new MqttProperties.StringProperty(propertyId.value, stringResult.value));
-                    break;
-                case 0x09: // Correlation Data => Binary Data
-                case 0x16: // Authentication Data
-                    final Result<byte[]> binaryDataResult = decodeByteArray(buffer);
-                    numberOfBytesConsumed += binaryDataResult.numberOfBytesConsumed;
-                    decodedProperties.add(new MqttProperties.BinaryProperty(propertyId.value, binaryDataResult.value));
-                    break;
-            }
-        }
-
-        return new Result<MqttProperties>(decodedProperties, numberOfBytesConsumed);
-    }
-
-    private static Result<MqttConnAckVariableHeader> decodeConnAckVariableHeader(ByteBuf buffer) {
-        final boolean sessionPresent = (buffer.readUnsignedByte() & 0x01) == 0x01;
-        byte returnCode = buffer.readByte();
-        final int numberOfBytesConsumed = 2;
-        final MqttConnAckVariableHeader mqttConnAckVariableHeader =
-                new MqttConnAckVariableHeader(MqttConnectReturnCode.valueOf(returnCode), sessionPresent);
-        return new Result<MqttConnAckVariableHeader>(mqttConnAckVariableHeader, numberOfBytesConsumed);
-    }
-
-    private static Result<MqttMessageIdVariableHeader> decodeMessageIdVariableHeader(ByteBuf buffer) {
-        final Result<Integer> messageId = decodeMessageId(buffer);
-        return new Result<MqttMessageIdVariableHeader>(
-                MqttMessageIdVariableHeader.from(messageId.value),
-                messageId.numberOfBytesConsumed);
-    }
-
-    private static Result<MqttPublishVariableHeader> decodePublishVariableHeader(
-            ByteBuf buffer,
-            MqttFixedHeader mqttFixedHeader) {
-        final Result<String> decodedTopic = decodeString(buffer);
-        if (!isValidPublishTopicName(decodedTopic.value)) {
-            throw new DecoderException("invalid publish topic name: " + decodedTopic.value + " (contains wildcards)");
-        }
-        int numberOfBytesConsumed = decodedTopic.numberOfBytesConsumed;
-
-        int messageId = -1;
-        if (mqttFixedHeader.qosLevel().value() > 0) {
-            final Result<Integer> decodedMessageId = decodeMessageId(buffer);
-            messageId = decodedMessageId.value;
-            numberOfBytesConsumed += decodedMessageId.numberOfBytesConsumed;
-        }
-        final MqttPublishVariableHeader mqttPublishVariableHeader =
-                new MqttPublishVariableHeader(decodedTopic.value, messageId);
-        return new Result<MqttPublishVariableHeader>(mqttPublishVariableHeader, numberOfBytesConsumed);
-    }
-
-    private static Result<Integer> decodeMessageId(ByteBuf buffer) {
-        final Result<Integer> messageId = decodeMsbLsb(buffer);
-        if (!isValidMessageId(messageId.value)) {
-            throw new DecoderException("invalid messageId: " + messageId.value);
-        }
-        return messageId;
-    }
+//    private static Result<MqttConnAckVariableHeader> decodeConnAckVariableHeader(ByteBuf buffer) {
+//        final boolean sessionPresent = (buffer.readUnsignedByte() & 0x01) == 0x01;
+//        byte returnCode = buffer.readByte();
+//        final int numberOfBytesConsumed = 2;
+//        final MqttConnAckVariableHeader mqttConnAckVariableHeader =
+//                new MqttConnAckVariableHeader(MqttConnectReturnCode.valueOf(returnCode), sessionPresent);
+//        return new Result<MqttConnAckVariableHeader>(mqttConnAckVariableHeader, numberOfBytesConsumed);
+//    }
+//
+//    private static Result<MqttMessageIdVariableHeader> decodeMessageIdVariableHeader(ByteBuf buffer) {
+//        final Result<Integer> messageId = decodeMessageId(buffer);
+//        return new Result<MqttMessageIdVariableHeader>(
+//                MqttMessageIdVariableHeader.from(messageId.value),
+//                messageId.numberOfBytesConsumed);
+//    }
+//
+//    private static Result<MqttPublishVariableHeader> decodePublishVariableHeader(
+//            ByteBuf buffer,
+//            MqttFixedHeader mqttFixedHeader) {
+//        final Result<String> decodedTopic = decodeString(buffer);
+//        if (!isValidPublishTopicName(decodedTopic.value)) {
+//            throw new DecoderException("invalid publish topic name: " + decodedTopic.value + " (contains wildcards)");
+//        }
+//        int numberOfBytesConsumed = decodedTopic.numberOfBytesConsumed;
+//
+//        int messageId = -1;
+//        if (mqttFixedHeader.qosLevel().value() > 0) {
+//            final Result<Integer> decodedMessageId = decodeMessageId(buffer);
+//            messageId = decodedMessageId.value;
+//            numberOfBytesConsumed += decodedMessageId.numberOfBytesConsumed;
+//        }
+//        final MqttPublishVariableHeader mqttPublishVariableHeader =
+//                new MqttPublishVariableHeader(decodedTopic.value, messageId);
+//        return new Result<MqttPublishVariableHeader>(mqttPublishVariableHeader, numberOfBytesConsumed);
+//    }
+//
+//    private static Result<Integer> decodeMessageId(ByteBuf buffer) {
+//        final Result<Integer> messageId = decodeMsbLsb(buffer);
+//        if (!isValidMessageId(messageId.value)) {
+//            throw new DecoderException("invalid messageId: " + messageId.value);
+//        }
+//        return messageId;
+//    }
 
     /**
      * Decodes the payload.
@@ -505,7 +506,7 @@ public final class MqttDecoder extends ReplayingDecoder<DecoderState> {
         return new Result<ByteBuf>(b, bytesRemainingInVariablePart);
     }
 
-    private static Result<String> decodeString(ByteBuf buffer) {
+    static Result<String> decodeString(ByteBuf buffer) {
         return decodeString(buffer, 0, Integer.MAX_VALUE);
     }
 
@@ -524,7 +525,7 @@ public final class MqttDecoder extends ReplayingDecoder<DecoderState> {
         return new Result<String>(s, numberOfBytesConsumed);
     }
 
-    private static Result<byte[]> decodeByteArray(ByteBuf buffer) {
+    static Result<byte[]> decodeByteArray(ByteBuf buffer) {
         final Result<Integer> decodedSize = decodeMsbLsb(buffer);
         int size = decodedSize.value;
         byte[] bytes = new byte[size];
@@ -532,7 +533,7 @@ public final class MqttDecoder extends ReplayingDecoder<DecoderState> {
         return new Result<byte[]>(bytes, decodedSize.numberOfBytesConsumed + size);
     }
 
-    private static Result<Integer> decodeMsbLsb(ByteBuf buffer) {
+    static Result<Integer> decodeMsbLsb(ByteBuf buffer) {
         return decodeMsbLsb(buffer, 0, 65535);
     }
 
@@ -547,7 +548,7 @@ public final class MqttDecoder extends ReplayingDecoder<DecoderState> {
         return new Result<Integer>(result, numberOfBytesConsumed);
     }
 
-    private static Result<Integer> decode4bytesInteger(ByteBuf buffer) {
+    static Result<Integer> decode4bytesInteger(ByteBuf buffer) {
         short msb = buffer.readUnsignedByte();
         short secondByte = buffer.readUnsignedByte();
         short thirdByte = buffer.readUnsignedByte();
@@ -558,10 +559,10 @@ public final class MqttDecoder extends ReplayingDecoder<DecoderState> {
         return new Result<Integer>(result, numberOfBytesConsumed);
     }
 
-    private static final class Result<T> {
+    static final class Result<T> {
 
-        private final T value;
-        private final int numberOfBytesConsumed;
+        final T value;
+        final int numberOfBytesConsumed;
 
         Result(T value, int numberOfBytesConsumed) {
             this.value = value;
