+++ codec-mqtt/src/main/java/io/netty/handler/codec/mqtt/VariableHeaderDecoderV5.java
@@ -1,0 +1,216 @@
+package io.netty.handler.codec.mqtt;
+
+import io.netty.buffer.ByteBuf;
+import io.netty.handler.codec.DecoderException;
+
+import static io.netty.handler.codec.mqtt.MqttCodecUtil.isValidMessageId;
+import static io.netty.handler.codec.mqtt.MqttCodecUtil.isValidPublishTopicName;
+import static io.netty.handler.codec.mqtt.MqttDecoder.*;
+
+/**
+ * Used to decode only the Variable Header part of MQTT packets, which is what differs between v3.1.1 and v5
+ */
+final class VariableHeaderDecoderV5  implements IVariableHeaderDecoder {
+    /**
+     * Decodes the variable header (if any)
+     * @param buffer the buffer to decode from
+     * @param mqttFixedHeader MqttFixedHeader of the same message
+     * @return the variable header
+     */
+    @Override
+    public Result<?> decodeVariableHeader(ByteBuf buffer, MqttFixedHeader mqttFixedHeader) {
+        switch (mqttFixedHeader.messageType()) {
+            case CONNECT:
+                return decodeConnectionVariableHeader(buffer);
+
+            case CONNACK:
+                return decodeConnAckVariableHeader(buffer);
+
+            case SUBSCRIBE:
+            case UNSUBSCRIBE:
+            case SUBACK:
+            case UNSUBACK:
+            case PUBACK:
+            case PUBREC:
+            case PUBCOMP:
+            case PUBREL:
+                return decodeMessageIdVariableHeader(buffer);
+
+            case PUBLISH:
+                return decodePublishVariableHeader(buffer, mqttFixedHeader);
+
+            case PINGREQ:
+            case PINGRESP:
+            case DISCONNECT:
+                // Empty variable header
+                return new Result<Object>(null, 0);
+        }
+        return new Result<Object>(null, 0); //should never reach here
+    }
+
+    private static Result<MqttConnectVariableHeader> decodeConnectionVariableHeader(ByteBuf buffer) {
+        final Result<String> protoString = decodeString(buffer);
+        int numberOfBytesConsumed = protoString.numberOfBytesConsumed;
+
+        final byte protocolLevel = buffer.readByte();
+        numberOfBytesConsumed += 1;
+
+        final MqttVersion mqttVersion = MqttVersion.fromProtocolNameAndLevel(protoString.value, protocolLevel);
+
+        final int b1 = buffer.readUnsignedByte();
+        numberOfBytesConsumed += 1;
+
+        final Result<Integer> keepAlive = decodeMsbLsb(buffer);
+        numberOfBytesConsumed += keepAlive.numberOfBytesConsumed;
+
+        final boolean hasUserName = (b1 & 0x80) == 0x80;
+        final boolean hasPassword = (b1 & 0x40) == 0x40;
+        final boolean willRetain = (b1 & 0x20) == 0x20;
+        final int willQos = (b1 & 0x18) >> 3;
+        final boolean willFlag = (b1 & 0x04) == 0x04;
+        final boolean cleanSession = (b1 & 0x02) == 0x02;
+        if (mqttVersion == MqttVersion.MQTT_3_1_1) {
+            final boolean zeroReservedFlag = (b1 & 0x01) == 0x0;
+            if (!zeroReservedFlag) {
+                // MQTT v3.1.1: The Server MUST validate that the reserved flag in the CONNECT Control Packet is
+                // set to zero and disconnect the Client if it is not zero.
+                // See http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc385349230
+                throw new DecoderException("non-zero reserved flag");
+            }
+        }
+
+        final Result<MqttProperties> properties = decodeProperties(buffer);
+        numberOfBytesConsumed += properties.numberOfBytesConsumed;
+
+        final MqttConnectVariableHeader mqttConnectVariableHeader = new MqttConnectVariableHeader(
+                mqttVersion.protocolName(),
+                mqttVersion.protocolLevel(),
+                hasUserName,
+                hasPassword,
+                willRetain,
+                willQos,
+                willFlag,
+                cleanSession,
+                keepAlive.value,
+                properties.value);
+        return new Result<MqttConnectVariableHeader>(mqttConnectVariableHeader, numberOfBytesConsumed);
+    }
+
+    private static Result<MqttConnAckVariableHeader> decodeConnAckVariableHeader(ByteBuf buffer) {
+        final boolean sessionPresent = (buffer.readUnsignedByte() & 0x01) == 0x01;
+        byte returnCode = buffer.readByte();
+        int numberOfBytesConsumed = 2;
+
+        final Result<MqttProperties> properties = decodeProperties(buffer);
+        numberOfBytesConsumed += properties.numberOfBytesConsumed;
+
+        final MqttConnAckVariableHeader mqttConnAckVariableHeader =
+                new MqttConnAckVariableHeader(MqttConnectReturnCode.valueOf(returnCode), sessionPresent,
+                        properties.value);
+
+        return new Result<MqttConnAckVariableHeader>(mqttConnAckVariableHeader, numberOfBytesConsumed);
+    }
+
+    private static Result<MqttMessageIdVariableHeader> decodeMessageIdVariableHeader(ByteBuf buffer) {
+        final Result<Integer> messageId = decodeMessageId(buffer);
+        return new Result<MqttMessageIdVariableHeader>(
+                MqttMessageIdVariableHeader.from(messageId.value),
+                messageId.numberOfBytesConsumed);
+    }
+
+    private static Result<MqttPublishVariableHeader> decodePublishVariableHeader(
+            ByteBuf buffer,
+            MqttFixedHeader mqttFixedHeader) {
+        final Result<String> decodedTopic = decodeString(buffer);
+        if (!isValidPublishTopicName(decodedTopic.value)) {
+            throw new DecoderException("invalid publish topic name: " + decodedTopic.value + " (contains wildcards)");
+        }
+        int numberOfBytesConsumed = decodedTopic.numberOfBytesConsumed;
+
+        int messageId = -1;
+        if (mqttFixedHeader.qosLevel().value() > 0) {
+            final Result<Integer> decodedMessageId = decodeMessageId(buffer);
+            messageId = decodedMessageId.value;
+            numberOfBytesConsumed += decodedMessageId.numberOfBytesConsumed;
+        }
+        final MqttPublishVariableHeader mqttPublishVariableHeader =
+                new MqttPublishVariableHeader(decodedTopic.value, messageId);
+        return new Result<MqttPublishVariableHeader>(mqttPublishVariableHeader, numberOfBytesConsumed);
+    }
+
+    private static Result<MqttProperties> decodeProperties(ByteBuf buffer) {
+        final Result<Integer> propertiesLength = decodeVariableByteInteger(buffer);
+        int totalPropertiesLength = propertiesLength.value;
+        int numberOfBytesConsumed = propertiesLength.numberOfBytesConsumed;
+
+        MqttProperties decodedProperties = new MqttProperties();
+        while (numberOfBytesConsumed < totalPropertiesLength) {
+            Result<Integer> propertyId = decodeVariableByteInteger(buffer);
+            numberOfBytesConsumed += propertyId.numberOfBytesConsumed;
+
+            switch (propertyId.value) {
+                case 0x01: // Payload Format Indicator => Byte
+                case 0x17: // Request Problem Information
+                case 0x19: // Request Response Information
+                case 0x24: // Maximum QoS
+                case 0x25: // Retain Available
+                case 0x28: // Wildcard Subscription Available
+                case 0x29: // Subscription Identifier Available
+                case 0x2A: // Shared Subscription Available
+                    final int b1 = buffer.readUnsignedByte();
+                    numberOfBytesConsumed ++;
+                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, b1));
+                    break;
+                case 0x13: // Server Keep Alive => Two Byte Integer
+                case 0x21: // Receive Maximum
+                case 0x22: // Topic Alias Maximum
+                case 0x23: // Topic Alias
+                    final Result<Integer> int2BytesResult = decodeMsbLsb(buffer);
+                    numberOfBytesConsumed += int2BytesResult.numberOfBytesConsumed;
+                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, int2BytesResult.value));
+                    break;
+                case 0x02: // Publication Expiry Interval => Four Byte Integer
+                case 0x11: // Session Expiry Interval
+                case 0x18: // Will Delay Interval
+                case 0x27: // Maximum Packet Size
+                    final Result<Integer> int4BytesResult = decode4bytesInteger(buffer);
+                    numberOfBytesConsumed += int4BytesResult.numberOfBytesConsumed;
+                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, int4BytesResult.value));
+                    break;
+                case 0x0B: // Subscription Identifier => Variable Byte Integer
+                    Result<Integer> vbIntegerResult = decodeVariableByteInteger(buffer);
+                    numberOfBytesConsumed += vbIntegerResult.numberOfBytesConsumed;
+                    decodedProperties.add(new MqttProperties.IntegerProperty(propertyId.value, vbIntegerResult.value));
+                    break;
+                case 0x03: // Content Type => UTF-8 Encoded String
+                case 0x08: // Response Topic
+                case 0x12: // Assigned Client Identifier
+                case 0x15: // Authentication Method
+                case 0x1A: // Response Information
+                case 0x1C: // Server Reference
+                case 0x1F: // Reason String
+                case 0x26: // User Property
+                    final Result<String> stringResult = decodeString(buffer);
+                    numberOfBytesConsumed += stringResult.numberOfBytesConsumed;
+                    decodedProperties.add(new MqttProperties.StringProperty(propertyId.value, stringResult.value));
+                    break;
+                case 0x09: // Correlation Data => Binary Data
+                case 0x16: // Authentication Data
+                    final Result<byte[]> binaryDataResult = decodeByteArray(buffer);
+                    numberOfBytesConsumed += binaryDataResult.numberOfBytesConsumed;
+                    decodedProperties.add(new MqttProperties.BinaryProperty(propertyId.value, binaryDataResult.value));
+                    break;
+            }
+        }
+
+        return new Result<MqttProperties>(decodedProperties, numberOfBytesConsumed);
+    }
+
+    private static Result<Integer> decodeMessageId(ByteBuf buffer) {
+        final Result<Integer> messageId = decodeMsbLsb(buffer);
+        if (!isValidMessageId(messageId.value)) {
+            throw new DecoderException("invalid messageId: " + messageId.value);
+        }
+        return messageId;
+    }
+}
